Track request analytics to private server as well as GA (#2)
To improve visitor data privacy, it would be cool to manage pseudo-personal data like IP addresses and browser fingerprints on servers we control.

To establish if this is feasible, this change enables logging to a private instance of Matomo analytics. In future, we can analyse the information gathered and shut down Google analytics if the new system suits our (admittedly basic) needs and doesn't cost too much to run.